requirements.txt
fastapi
uvicorn
sqlalchemy
jinja2
ddddocr
passlib[bcrypt]
python-jose 